+++ configs/file.js
@@ -1,0 +1,67 @@
+'use strict';
+
+var path = require('path');
+var pkg = require(path.resolve('package'));
+var modules = [].concat(Object.keys(pkg.dependencies || {}),
+                        Object.keys(pkg.devDependencies || {}),
+                        Object.keys(pkg.peerDependencies || {}));
+var extensions = exports.extensions = ['.js'];
+var glob = ['js'];
+var loaders = exports.loaders = [];
+var aliases = exports.aliases = {};
+
+// JS with Traceur
+if(moduleExists('traceur-loader')) {
+  loaders.push({
+    test: /^(?!.*(bower_components|node_modules))+.+\.js$/,
+    loaders: ['traceur-loader?experimental']
+  });
+  aliases['traceur-runtime'] = require('traceur-loader').runtime;
+}
+
+// JSX
+if(moduleExists('jsx-loader')) {
+  extensions.push('.jsx');
+  loaders.push('jsx');
+  loaders.push({
+    test: /\.jsx$/,
+    loaders: ['jsx-loader?harmony&insertPragma=React.DOM']
+  });
+}
+
+// CoffeeScript
+if(moduleExists('coffee-loader')) {
+  extensions.push('.coffee', '.coffee.md', '.litcoffee');
+  glob.push('coffee', 'coffee.md', 'litcoffee');
+  loaders.push({
+    test: /\.(coffee|coffee\.md|litcoffee)$/,
+    loaders: ['coffee-loader']
+  });
+}
+
+// HTML
+if(moduleExists('html-loader')) {
+  extensions.push('.html');
+  loaders.push({
+    test: /\.html$/,
+    loaders: ['html-loader']
+  });
+}
+
+// Build glob
+if(glob.length > 1) {
+  exports.glob = '*.{' + glob.join(',') + '}';
+}
+else if(glob.length === 1) {
+  exports.glob = '*.' + glob[0];
+}
+
+function moduleExists(mod) {
+  if(~modules.indexOf(mod)) {
+    try {
+      require.resolve(mod);
+      return true;
+    }
+    catch(e) {}
+  }
+}